Refactor workflow tests to use executors. Remove old runners and publishing related bootstrap files
