+++ .env.example
@@ -1,2 +1,1 @@
 GT_CLOUD_API_KEY=<your_api_key>
-GTN_USE_WEBSOCKETS=false
