--- src/griptape_nodes/bootstrap/post_build_install_script.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-python register_libraries_script.py
