--- src/griptape_nodes/bootstrap/pre_build_install_script.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-apt-get install -y curl
-curl -LsSf https://astral.sh/uv/install.sh | sh
